cloudfunction/pubsub_to_gcs/source/requirements.txt
google-api-python-client==1.12.5
google-api-core==1.23.0
google-cloud-storage==1.32.0
